Reach stopping point on style work to refactor DB connections.
